+++ myclimbz/templates/form_sort_videos.html
@@ -21,9 +21,8 @@
         <div style="display:none;"> {{ form.videos_sorted() }} </div>
         <input type="submit" class="btn btn-success" name="btn" value="Continue">
          <a href="{{ url_for('home.cancel_form') }}" class="btn btn-danger ms-2">Cancel Upload</a>
-         <!-- TODO: cancel upload not working -->
     </form>
-
+    
 </div>
 
 <script src="https://cdnjs.cloudflare.com/ajax/libs/Sortable/1.15.0/Sortable.min.js"></script>
